+++ src/app/about/page.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import About from "./About";
+
+export default function page() {
+    return <About />;
+}
