+++ src/app/contact/page.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import Contact from "./Contact";
+
+export default function page() {
+    return <Contact />;
+}
