+++ src/app/projects/page.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import Projects from "./Projects";
+
+export default function page() {
+    return <Projects />;
+}
