+++ src/app/skills/page.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import Skills from "./Skills";
+
+export default function page() {
+    return <Skills />;
+}
